fix binarySearch in DataLoader

+++ src/cse291/lsmdb/io/interfaces/StringHasher.java
@@ -4,5 +4,5 @@ package cse291.lsmdb.io.interfaces;
  * Created by musteryu on 2017/5/29.
  */
 public interface StringHasher {
-    int hash(String key);
+    long hash(String key);
 }
